zoom_clone/requirements.txt
Django>=4.2,<5.0
gunicorn>=20.1.0
python-decouple>=3.8,<4.0
dj-database-url>=1.0.0
psycopg2-binary>=2.9.6
pytz>=2023.3
sqlparse>=0.4.4
channels[daphne]>=4.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
whitenoise>=6.4.0
daphne>=4.0.0